20250423
Several changes to the functions. I've now tied the call_id to the agency, cleaned up the names to remove Mr and Mrs, changed the distributions and added time stamps for events based on the elapsed times.  Still need to clean parts of that up, but progress is being made.